Refactor UUID, username, and owner checking
This system really needs a total redesign to account for the fact that Mojang's usernames don't all appear under their web api, as a result of having what I assume are Microsoft accounts. This caused a NullPointerException to appear due to a blank HTTP GET response for an invalid username.