Change how damage doubling is done in Light/Dark abilities
Previously, damage was doubled before armor reductions and after buffs, thus making the final damage more than doubled.
This is now replaced by a final damage calculation based on how the game actually determines it.
All code that used to calculate if something was fatal needs to also be corrected from now on, as previous code did not check for absorption hearts.